tests, progress, call signatures

+++ Tests/nt.Client.Integration.Tests/Features/WebThree/GetNftTypes/FetchNftTypesTests.cs
@@ -29,7 +29,7 @@
       WebThreeState ActionResponse = await Mediator.Send(Action);
 
       ActionResponse.ShouldNotBe(null);
-      ActionResponse.TotalNftTypes.ShouldBeGreaterThan(2);
+      ActionResponse.TotalNftTypes.ShouldBeGreaterThan((uint)2);
 
     }
 
@@ -39,7 +39,7 @@
 
       WebThreeState response = await Mediator.Send(fetchNftTypes);
 
-      response.TotalNftTypes.ShouldBeGreaterThan(2);
+      response.TotalNftTypes.ShouldBeGreaterThan((uint)2);
     }
 
   }
